desafio03

+++ Aula15/desafio03.py
@@ -9,7 +9,7 @@ dados['idade'] = date.today().year - dados['idade']
 
 if dados['ctps'] != 0:
     dados['contratação'] = int(input('Ano de contratação: '))
-    dados['salário'] = int(input('Salário: R$ '))
+    dados['salário'] = float(input('Salário: R$ '))
     dados['aposentadoria'] = (35 - (date.today().year - dados['contratação'])) + dados['idade']
     print('=' * 30)
